fix setup script

+++ src/lib/cleanProject.ts
@@ -1,7 +1,7 @@
 import fs from 'node:fs'
 import { printLog } from './logger'
 
-const filesToDelete = ['dist', '.next', '*.g.ts', 'node_modules', '.svelte-kit']
+const filesToDelete = ['dist', '.next', '*.g.ts', '.svelte-kit']
 
 export const cleanProject = (dir: string = process.cwd()) => {
 	if (dir === process.cwd()) printLog('CLEAN PROJECT', 'warn')
@@ -11,6 +11,7 @@ export const cleanProject = (dir: string = process.cwd()) => {
 	for (const path of paths) {
 		try {
 			if (path.includes('.git')) continue
+			if (path.includes('node_modules')) continue
 
 			const filePath = `${dir}/${path}`
 			if (filesToDelete.includes(path)) fs.rmSync(filePath, { force: true, recursive: true })
